refactor: consolidate per-command files back into root.go / main.go
Merge the over-split per-command files back into single files per package:

- internal/cli/renv/root.go: absorbs registry.go, resolve.go, exec.go,
  yaml_cmd.go, shell.go, status.go, cache.go, unload.go, watch.go
- internal/cli/kctx/root.go: absorbs registry.go, load.go, switch.go,
  shell.go, status.go, cache.go, unload.go, watch.go
- cmd/envoke/main.go: absorbs resolve.go, unload.go, shell.go, cache_watch.go

Platform-specific detach_unix.go and detach_windows.go remain separate.
Imports deduplicated; all functions, constants, and helpers preserved.
Build and tests verified passing.
